ref: move headings, links and props to atom components
