+++ src/components/atoms/headings/H2HeadingSubtitle.tsx
@@ -1,8 +1,8 @@
-import type { DefaultComponentProps } from '../../types/componentProps';
+import type { DefaultAtomProps } from '../AtomProps';
 
 export default function H1HeadingSubtitle({
   children,
   className = 'text-center mb-4 text-tertiary',
-}: DefaultComponentProps) {
+}: DefaultAtomProps) {
   return <p className={className}>{children}</p>;
 }
