+++ src/components/atoms/headings/H3Heading.tsx
@@ -1,8 +1,8 @@
-import type { DefaultComponentProps } from '../../types/componentProps';
+import type { DefaultAtomProps } from '../AtomProps';
 
 export default function H3Heading({
   children,
   className = 'text-center mb-4 text-2xl text-content',
-}: DefaultComponentProps) {
+}: DefaultAtomProps) {
   return <h3 className={className}>{children}</h3>;
 }
