+++ src/pages/NotFound.tsx
@@ -1,4 +1,4 @@
-import H1Heading from '../components/headings/H1Heading';
+import H1Heading from '../components/atoms/headings/H1Heading';
 
 export default function NotFoundPage() {
   return <H1Heading>Not Found!</H1Heading>;
